requirements.txt
openai
langchain
langchain-community
langchain_openai
faiss-cpu
python-dotenv
streamlit
fastapi
uvicorn
python-multipart
pypdf
tiktoken